Added error checking throughout program. Commented retrieve module.

+++ MP1Tweets.py
@@ -17,13 +17,22 @@ def grab(handle):
 
 	#Authorize with API.
 	print("Authorizing with Twitter...")
-	auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
-	auth.set_access_token(access_key, access_secret)
-	api = tweepy.API(auth)
+	try:
+		auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
+		auth.set_access_token(access_key, access_secret)
+		api = tweepy.API(auth)
+	except:
+		return 0
 
-	print("Collecting tweets...")
+	print("Checking twitter handle...")
 
 	#Add check for handle existence.
+	try:
+		api.get_user(handle)
+	except:
+		return 1
+
+	print("Collecting tweets...")
 
 	#Stage variable for storing urls.
 	imageurls = []
@@ -35,7 +44,7 @@ def grab(handle):
 
 	#Loop through tweets until no more tweets or 500 tweets have been investigated.
 	#Limit for status request is 900.
-	while len(newtweet) > 0:
+	while (len(newtweet) > 0 and tweetnum < 501):
 		print("Tweet found: ", newtweet[0].id)
 		#Check if the current tweet is a retweet.
 		if hasattr(newtweet[0], 'retweeted_status'):
@@ -52,9 +61,8 @@ def grab(handle):
 		newid = newtweet[0].id - 1
 		#Get new tweet.
 		newtweet = api.user_timeline(screen_name = handle,count=1,max_id=newid)
-		#Check to see if tweet limit is reached.
-		if tweetnum > 500
-			break
+		#Add to tweet number.
+		tweetnum = tweetnum +1
 
 	print("Max tweets collected.")
 
@@ -62,26 +70,54 @@ def grab(handle):
 
 
 def retrieve(imageurls):
-	print("TweetParse Called")
+	print("Checking ./ImageDump folder...")
 
+	#Check if folder already exists from previous instance of this program.
 	if os.path.isdir("./ImageDump"):
-		#Insert warning and confirmation before deletion.
-		shutil.rmtree("./ImageDump")
+		#Confirm deletion.
+		print("./ImageDump exists; to proceed it must be deleted which will erase all contents.")
+		delconf = str(input("Confirm deletion of ./ImageDump y/n: "))
+		while True:
+			if (delconf is "y" or delconf is "n" or delconf is "Y" or delconf is "N"):
+				break
+			else:
+				delconf = str(input("Please input y or n: "))
+		if (delconf is "y" or delconf is "Y"):
+			shutil.rmtree("./ImageDump")
+			print("Folder deleted, creating new one...")
+		else:
+			return(1)
+	else:
+		print("Folder doesn't exist, creating...")
 
+	#Create folder and move to it.
 	os.mkdir("ImageDump")
 	os.chdir("./ImageDump")
 
+	#Stage values for image naming.
 	imagenum = 1
+	imagestr = "000"
 
 	for image in imageurls:
+		#Check for how many zeros to add in front of number.
+		#Since I limit the tweet grabber to 500 tweets, 
+		#no more than that number of images is expected.
 		if imagenum < 10:
-			urllib.request.urlretrieve(image, "000%s.jpg" % (imagenum))
+			imagestr = "00%s.jpg" % (imagenum)
 		elif imagenum < 100:
-			urllib.request.urlretrieve(image, "00%s.jpg" % (imagenum))
+			imagestr = "0%s.jpg" % (imagenum)
 		elif imagenum < 1000:
-			urllib.request.urlretrieve(image, "0%s.jpg" % (imagenum))
-		elif imagenum < 10000:
-			urllib.request.urlretrieve(image, "%s.jpg" % (imagenum))
+			imagestr = "%s.jpg" % (imagenum)
+		else:
+			print("Too many images, stopping...")
+			break
+		#Try to download image, print status of each image retrieval.
+		try:
+			urllib.request.urlretrieve(image, imagestr)
+		except:
+			print("Image: %s error, cannot retrieve..." % (imagestr))
+		else:
+			print("Image: %s retrieved..." % (imagestr))
 		imagenum = imagenum + 1
